Fix size argument in cross-over mutation
If the length of fittest genes gets very low, it is possible that the
initial gene length parameter is larger than the available gene pool,
which causes an error. This bug has been fixed by using the minimum of
the gene pool size and the initila gene length as the size parameter.