Make max and min lengths of mutation variable
Previously, there was no way to call the `set_min/max_length` methods
from the other modules. The min and max length are now instance
variables of the respective mutation classes and can be specified in the
constructor.

Added a new format option for specifying the mutations in the input file
as a dictionary. This format is currently optional.