Fix NaN handling in MultiMetric
The genetic algorith produced too many NaN values when using larger
member sizes. This is now circumvented by using numpy's `nanmean`
function to ignore the NaN values when computing the mean.

A warning is thrown if there are too many NaN values in the metric array
before averaging and the test point locations where the NaN values occur
are tracked and can be seen with the log level DEBUG.