Add option to revert actinide reduction
The actinide reduction can be reverted after filtering so that the data
written to the csv file reflects the concentrations in the input.

If this data is used as training data, it reflects the quantities of
each nuclide in the SNF and can be used to predict U and Pu content.